feat: typescript support
