+++ .eslintrc
@@ -1,0 +1,7 @@
+{
+  "extends": "next/core-web-vitals",
+  "rules": {
+    "@next/next/no-img-element": "off",
+    "import/no-anonymous-default-export": "off"
+  }
+}
