+++ pages/api/generate.ts
@@ -1,13 +1,14 @@
-import { Configuration, OpenAIApi } from "openai";
+import { NextApiRequest, NextApiResponse } from 'next';
+import { Configuration, OpenAIApi } from 'openai';
 
 const configuration = new Configuration({
   apiKey: process.env.OPENAI_API_KEY,
 });
 const openai = new OpenAIApi(configuration);
 
-export default async function (req, res) {
+export default async function (req: NextApiRequest, res: NextApiResponse) {
   const completion = await openai.createCompletion({
-    model: "text-davinci-002",
+    model: 'text-davinci-002',
     prompt: generatePrompt(req.body.animal),
     temperature: 0.6,
   });
