+++ pages/index.tsx
@@ -1,0 +1,50 @@
+import Head from 'next/head';
+import { FormEvent, useCallback, useState } from 'react';
+import styles from './index.module.css';
+
+export default function Home() {
+  const [animalInput, setAnimalInput] = useState('');
+  const [result, setResult] = useState();
+
+  const onSubmit = useCallback(
+    async (event: FormEvent<HTMLFormElement>) => {
+      event.preventDefault();
+      const response = await fetch('/api/generate', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ animal: animalInput }),
+      });
+      const data = await response.json();
+      setResult(data.result);
+      setAnimalInput('');
+    },
+    [animalInput],
+  );
+
+  return (
+    <div>
+      <Head>
+        <title>OpenAI Quickstart</title>
+        <link rel="icon" href="/dog.png" />
+      </Head>
+
+      <main className={styles.main}>
+        <img src="/dog.png" alt="dog" className={styles.icon} />
+        <h3>Name my pet</h3>
+        <form onSubmit={onSubmit}>
+          <input
+            type="text"
+            name="animal"
+            placeholder="Enter an animal"
+            value={animalInput}
+            onChange={(e) => setAnimalInput(e.target.value)}
+          />
+          <input type="submit" value="Generate names" />
+        </form>
+        <div className={styles.result}>{result}</div>
+      </main>
+    </div>
+  );
+}
